Refactor: DI 모듈 이동 & 오류 수정
