+++ app/src/main/java/com/kappzzang/jeongsan/di/UseCaseModule.kt
@@ -3,13 +3,10 @@ package com.kappzzang.jeongsan.di
 import com.kappzzang.jeongsan.repository.ExpenseDetailRepository
 import com.kappzzang.jeongsan.repository.ExpenseRepository
 import com.kappzzang.jeongsan.repository.GroupInfoRepository
-import com.kappzzang.jeongsan.repository.KakaoAuthenticationRepository
 import com.kappzzang.jeongsan.repository.MemberRepository
 import com.kappzzang.jeongsan.repository.ReceiptRepository
 import com.kappzzang.jeongsan.repository.TransferRepository
 import com.kappzzang.jeongsan.repository.UserInfoRepository
-import com.kappzzang.jeongsan.usecase.AuthenticateWithKakaoUseCase
-import com.kappzzang.jeongsan.usecase.AuthorizeWithKakaoUseCase
 import com.kappzzang.jeongsan.usecase.EditExpenseDetailUseCase
 import com.kappzzang.jeongsan.usecase.GetDoneGroupUseCase
 import com.kappzzang.jeongsan.usecase.GetExpenseDetailUseCase
@@ -17,12 +14,9 @@ import com.kappzzang.jeongsan.usecase.GetExpenseUseCase
 import com.kappzzang.jeongsan.usecase.GetInviteInfoUseCase
 import com.kappzzang.jeongsan.usecase.GetProgressingGroupUseCase
 import com.kappzzang.jeongsan.usecase.GetTransferInfoUseCase
-import com.kappzzang.jeongsan.usecase.GetUserInfoUseCase
-import com.kappzzang.jeongsan.usecase.RegisterWithKakaoUseCase
 import com.kappzzang.jeongsan.usecase.SendTransferMessageUseCase
 import com.kappzzang.jeongsan.usecase.UploadExpenseUseCase
 import com.kappzzang.jeongsan.usecase.UploadGroupInfoUseCase
-import com.kappzzang.jeongsan.util.AuthenticationRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -39,10 +33,6 @@ object UseCaseModule {
     @Provides
     fun provideGetDoneGroupUseCase(groupInfoRepository: GroupInfoRepository) =
         GetDoneGroupUseCase(groupInfoRepository)
-
-    @Provides
-    fun provideGetUserInfoUseCase(userInfoRepository: UserInfoRepository) =
-        GetUserInfoUseCase(userInfoRepository)
 
     @Provides
     fun provideGetInviteInfoUseCase(memberRepository: MemberRepository): GetInviteInfoUseCase =
@@ -65,22 +55,10 @@ object UseCaseModule {
         UploadExpenseUseCase(receiptRepository)
 
     @Provides
-    fun provideAuthenticateWithKakaoUseCase(
-        authenticationRepository: AuthenticationRepository,
-        kakaoAuthenticationRepository: KakaoAuthenticationRepository
-    ) = AuthenticateWithKakaoUseCase(authenticationRepository, kakaoAuthenticationRepository)
-
-    @Provides
-    fun provideAuthorizeWithKakaoUseCase(authenticationRepository: AuthenticationRepository) =
-        AuthorizeWithKakaoUseCase(authenticationRepository)
-
-    @Provides
-    fun registerWithKakaoUseCase() = RegisterWithKakaoUseCase()
-
-    @Provides
     fun provideUploadGroupInfoUseCase(groupInfoRepository: GroupInfoRepository) =
         UploadGroupInfoUseCase(groupInfoRepository)
 
+    @Provides
     fun provideGetTransferInfoUseCase(transferRepository: TransferRepository) =
         GetTransferInfoUseCase(transferRepository)
